+++ modules/home/forumPosts.json
@@ -1,0 +1,16 @@
+[
+  {
+    "title": "Clean money initiative",
+    "summary": "MakerDAO is in the position to use the most powerful tool ever invented to fight the most challenging problem ever faced - all it takes is the will of a community.",
+    "username": "Rune",
+    "link": "https://forum.makerdao.com/t/the-case-for-clean-money/10684",
+    "image": "/home/participation/clean-money.png"
+  },
+  {
+    "title": "Improving governance participation rate",
+    "summary": "With this post I’d like to zoom into one particular problem that Maker governance is facing today: Low governance participation rate.",
+    "username": "0xdeniz",
+    "link": "https://forum.makerdao.com/t/improving-governance-participation-rate-at-makerdao/11706",
+    "image": "/home/participation/improve-participation.png"
+  }
+]
